restructure files again. make api directory a module
