+++ .gitignore
@@ -1,1 +1,3 @@
 *__pycache__*
+*.egg-info
+build/
