+++ requirements.txt
@@ -1,1 +1,2 @@
 requests==2.32.3
+SQLAlchemy==2.0.32
